requirements.txt
psycopg2-binary==2.9.9
pyyaml==6.0.1
jsonschema==4.23.0
requests==2.32.3
